refactor: ledger
